lab6/requirements.txt
Flask==3.1.1
Flask-RESTful==0.3.10
marshmallow==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
flasgger==0.9.7.1